BACKEND/requirements.txt
requests==2.28.1
exif==1.3.5
Flask==2.0.1
Pillow==9.3.0
numpy==1.20.2
torch==1.9.0
torchvision==0.10.0
ipython==8.5.0
djitellopy==2.4.0
